feat(typer): add integration helpers for context injection
