+++ src/diwire/integrations/typer/__init__.py
@@ -1,0 +1,6 @@
+from diwire._internal.integrations.typer import add_typer_context, get_typer_context
+
+__all__ = [
+    "add_typer_context",
+    "get_typer_context",
+]
