+++ tests/e2e/typer/setup/app.py
@@ -1,31 +1,16 @@
 from __future__ import annotations
 
-from typing import cast
-
-import click
 import typer
 
 from diwire import Container, Injected, Lifetime, Scope, resolver_context
-from diwire.exceptions import DIWireIntegrationError
+from diwire.integrations.typer import add_typer_context
 from tests.e2e.typer.setup.services import ContextInfoService
 
 app = typer.Typer()
 
 
-def get_typer_context() -> typer.Context:
-    try:
-        return cast("typer.Context", click.get_current_context())
-    except RuntimeError as error:
-        msg = "Typer context not available. Ensure the command runs inside a Typer application."
-        raise DIWireIntegrationError(msg) from error
-
-
 container = Container()
-container.add_factory(
-    get_typer_context,
-    provides=typer.Context,
-    lifetime=Lifetime.TRANSIENT,
-)
+add_typer_context(container)
 container.add(
     ContextInfoService,
     scope=Scope.REQUEST,
